requirements.txt
-e git+git@github.com:divio/django-cms.git@release/3.5.x#egg=django-cms
# django-cms>=3.5,<3.6

djangocms-admin-style>=1.2,<1.3
django-treebeard>=4.0,<5.0

djangocms-text-ckeditor>=3.6.0
djangocms-link>=2.1
djangocms-googlemap>=1.1
djangocms-snippet>=2.0
djangocms-video>=2.0
djangocms-file>=2.0,<3.0
djangocms-picture>=2.0,<3.0
easy_thumbnails
django-filer>=1.3
Django<2.0
pytz
django-classy-tags>=0.7
html5lib>=0.999999,<0.99999999
Pillow>=3.0
django-sekizai>=0.9
six

django-settings-export
djangocms-icon
djangocms-bootstrap4
psycopg2-binary
django-parler
aldryn-translation-tools
django-admin-sortable2
aldryn-apphooks-config

djangocms-maps
django-dynamic-preferences
djangorestframework
django-filter
markdown

raven

https://control.divio.com/api/v1/apps/serve/djangocms-modules/0.0.4/ca71b210-4490-4a10-86e1-d023f93e6a5b/djangocms-modules-0.0.4.tar.gz#egg=djangocms-modules==0.0.4

django-google-tag-manager
